rename entrance to item
